+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "silly-bot",
-  "version": "1.0.11",
+  "version": "1.0.12",
   "description": "A little discord bot project",
   "main": "index.js",
   "scripts": {
