+++ src/database/index.ts
@@ -1,0 +1,12 @@
+import { setup } from "./mongo";
+import { points } from "./points";
+import { guild } from "./guild";
+import { counting } from "./counting";
+
+export const db = {
+    setup,
+    points,
+    guild,
+    counting,
+};
+
